+++ scriptconfig/dataconfig.py
@@ -240,6 +240,10 @@ class DataConfig(Config, metaclass=MetaDataConfig):
 
     def __init__(self, *args, **kwargs):
         "__autogenerateme__"
+        # Private internal hack to prevent __post_init__ from being called
+        # if we are immediately going to load and call it again.
+        _dont_call_post_init = kwargs.pop('_dont_call_post_init', False)
+
         self._data = None
         self._default = OrderedDict()
         if getattr(self, '__default__', None):
@@ -257,7 +261,8 @@ class DataConfig(Config, metaclass=MetaDataConfig):
         self._default.update(new_defaults)
         self._data = self._default.copy()
         self._enable_setattr = True
-        self.__post_init__()
+        if not _dont_call_post_init:
+            self.__post_init__()
 
     def __getattr__(self, key):
         # Note: attributes that mirror the public API will be suppressed